requirements.txt
albumentations==0.5.2
click==8.0.4
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
imgaug==0.4.0
kiwisolver==1.3.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.19.5
opencv-python==4.5.2.52
opencv-python-headless==4.5.2.52
pandas==1.2.4
Pillow==8.2.0
psutil==5.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.18.1
scipy==1.6.3
Shapely==1.7.1
six==1.16.0
tifffile==2021.4.8
tqdm==4.60.0
xxhash==3.0.0
